requirements.txt
twilio==6.25.2
PyQt5==5.12.1
PyYAML==5.1
